+++ e2e/cypress/support/commands.js
@@ -404,7 +404,7 @@ Cypress.Commands.add('hideTooltip', { prevSubject: 'optional' }, () => {
 });
 
 Cypress.Commands.add('waitForVideoToLoad', { prevSubject: 'optional' }, () => {
-  cy.get('[data-loaded=true]', { timeout: 15000 });
+  cy.get('#rich-content-viewer [data-loaded=true]', { timeout: 15000 });
 });
 
 // disable screenshots in debug mode. So there is no diffrence to ci.