+++ e2e/tests/rendering.e2e.js
@@ -1,14 +1,6 @@
-/*global cy*/
+/*global cy Cypress*/
 import { fixtures } from './constants';
 import { DEFAULT_DESKTOP_BROWSERS, DEFAULT_MOBILE_BROWSERS } from '../tests/constants';
-
-// function testViewerAndEditorAreEqual() {
-//   function getTitle(test = Cypress.mocha.getRunner().test) {
-//     return (test.parent && test.parent.title ? `${getTitle(test.parent)} > ` : '') + test.title;
-//   }
-//   cy.get('.DraftEditor-root').matchImageSnapshot(getTitle());
-//   cy.get('#root > div:nth-child(2) > div').matchImageSnapshot(getTitle());
-// }
 
 const testFixture = fixture =>
   it(`render ${fixture}`, function() {